Refactor hf model logic
