+++ olive/passes/onnx/conversion.py
@@ -13,7 +13,7 @@ from olive.common.config_utils import validate_config
 from olive.common.utils import tensor_data_to_device
 from olive.hardware.accelerator import AcceleratorSpec
 from olive.model import CompositeOnnxModel, ONNXModel, PyTorchModel
-from olive.model.hf.hf_utils import get_hf_model_io_config
+from olive.model.hf_utils import get_hf_model_io_config
 from olive.model.model_config import IOConfig
 from olive.passes import Pass
 from olive.passes.onnx.common import get_external_data_config, model_proto_to_olive_model
@@ -64,7 +64,7 @@ class OnnxConversion(Pass):
         dummy_inputs = model.get_dummy_inputs()
 
         # get input and output names, and dynamic axes
-        if model.hf_config and not model.hf_config.use_custom_implementation:
+        if model.hf_config and not model.hf_config.components:
             io_config = get_hf_model_io_config(
                 model.hf_config.model_name, model.hf_config.task, model.hf_config.feature
             )
